+++ CHANGELOG.md
@@ -7,6 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2020-02-29
+
+### Changed
+
+- `QueryRequest` now extends `Request`
+- update `req_builder` to use new URI fragments
+- add `QueryReqeust.copyWith`
+- add `clientError` object to `Query` Widget
+- `req_builder` no longer assigns a unique `queryId` if none is provided
+
+### Removed
+
+- Removed Options objects
+
 ## [0.2.1] - 2020-02-22
 
 ### Changed